% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MouldResume.R
\name{mdlMouldResume_mould_delete}
\alias{mdlMouldResume_mould_delete}
\title{模具删除}
\usage{
mdlMouldResume_mould_delete(dms_token, MouldNumber)
}
\arguments{
\item{MouldNumber}{}
}
\value{

}
\description{
模具删除
}
\examples{
mdlMouldResume_mould_delete()
}
