% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MouldResume.R
\name{mdlMouldResume_mould_viewall}
\alias{mdlMouldResume_mould_viewall}
\title{模具查询}
\usage{
mdlMouldResume_mould_viewall(dms_token, MouldNumber)
}
\arguments{
\item{MouldNumber}{}
}
\value{

}
\description{
模具查询
}
\examples{
mdlMouldResume_mould_viewall()
}
