% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MouldResume.R
\name{mdlMouldResume_previewall}
\alias{mdlMouldResume_previewall}
\title{模板履历预处理查询}
\usage{
mdlMouldResume_previewall(dms_token, FYEAR, FMONTH)
}
\value{

}
\description{
模板履历预处理查询
}
\examples{
mdlMouldResume_previewall()
}
