% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MouldResume.R
\name{mdlMouldResume_view}
\alias{mdlMouldResume_view}
\title{模板履历表查询}
\usage{
mdlMouldResume_view(token, FMoldNumber)
}
\arguments{
\item{FMoldNumber}{}
}
\value{

}
\description{
模板履历表查询
}
\examples{
mdlMouldResume_view(token,FMoldNumber)
}
