% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MouldResume.R
\name{MouldResume_delete}
\alias{MouldResume_delete}
\title{模板履历表删除}
\usage{
MouldResume_delete(token, FMoldNumber, FOutboundDate)
}
\arguments{
\item{FOutboundDate}{}
}
\value{

}
\description{
模板履历表删除
}
\examples{
MouldResume_delete()
}
