% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MouldResume.R
\name{MouldResume_update_preview}
\alias{MouldResume_update_preview}
\title{模具履历表预览}
\usage{
MouldResume_update_preview(token, file_name)
}
\arguments{
\item{file_name}{}
}
\value{

}
\description{
模具履历表预览
}
\examples{
MouldResume_update_preview()
}
