% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MouldResume.R
\name{MouldResume_update_upload}
\alias{MouldResume_update_upload}
\title{模具履历表上传}
\usage{
MouldResume_update_upload(dms_token, file_name)
}
\arguments{
\item{file_name}{}
}
\value{

}
\description{
模具履历表上传
}
\examples{
MouldResume_update_upload(dms_token,file_name)
}
