% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MouldResume.R
\name{MouldResume_upload}
\alias{MouldResume_upload}
\title{模具信息表上传}
\usage{
MouldResume_upload(token, file_name)
}
\arguments{
\item{file_name}{}
}
\value{

}
\description{
模具信息表上传
}
\examples{
MouldResume_upload(token,file_name)
}
