% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/productdaily.R
\name{productdaily_delete}
\alias{productdaily_delete}
\title{计时计件工资表删除}
\usage{
productdaily_delete(dms_token, FProductLots)
}
\arguments{
\item{FProductLots}{}
}
\value{

}
\description{
计时计件工资表删除
}
\examples{
productdaily_delete()
}
