% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/productdaily.R
\name{productdaily_erpsync}
\alias{productdaily_erpsync}
\title{同步erp数据到数据中台}
\usage{
productdaily_erpsync(erp_token, dms_token, FYEAR, FMONTH)
}
\arguments{
\item{FMONTH}{}
}
\value{

}
\description{
同步erp数据到数据中台
}
\examples{
productdaily_erpsync()
}
