% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/productdaily.R
\name{productdaily_upload}
\alias{productdaily_upload}
\title{日报表上传}
\usage{
productdaily_upload(dms_token, file_name)
}
\arguments{
\item{token}{}
}
\value{

}
\description{
日报表上传
}
\examples{
productdaily_upload()
}
