% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/productdaily.R
\name{productdaily_view}
\alias{productdaily_view}
\title{查询数据}
\usage{
productdaily_view(dms_token, FProductLots)
}
\arguments{
\item{FProductLots}{}
}
\value{

}
\description{
查询数据
}
\examples{
productdaily_view()
}
