% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MouldResume.R
\name{MouldResume_preview}
\alias{MouldResume_preview}
\title{模板履历表预览}
\usage{
MouldResume_preview(token, file_name)
}
\arguments{
\item{file_name}{}
}
\value{

}
\description{
模板履历表预览
}
\examples{
MouldResume_preview()
}
